Fix wallet directory initialization
util/system datadir code should not care about wallets or create any wallet paths
If -walletdir is specified, wallet init code should use that directory
If -walletdir is specified and path is not a directory, return a fatal error
If -walletdir is not specified, use <datadir>/wallets if it is directory.
If -walletdir is not specified and <datadir>/wallets does not exist:
<datadir>/wallets wallet should call ListWalletDir ListDatabases on <datadir>
If list is empty wallet should create <datadir>/wallets as a directory,
otherwise it should create it as a symlink pointing to .
discussion at https://github.com/bitcoin/bitcoin/issues/16220

+++ test/functional/test_framework/test_framework.py
@@ -841,8 +841,8 @@ class BitcoinTestFramework(metaclass=BitcoinTestMetaClass):
 
             def cache_path(*paths):
                 return os.path.join(cache_node_dir, self.chain, *paths)
-
-            os.rmdir(cache_path('wallets'))  # Remove empty wallets dir
+            if os.path.exists(cache_path('wallets')):
+                os.rmdir(cache_path('wallets'))  # Remove empty wallets dir
             for entry in os.listdir(cache_path()):
                 if entry not in ['chainstate', 'blocks', 'indexes']:  # Only indexes, chainstate and blocks folders
                     os.remove(cache_path(entry))
